requirements.txt
streamlit
polars
plotly
dash
pandas
google.generativeai
openpyxl